Refactor updateNextPrayer & remove calcTimeDelta
To support the functionality of calculating the amount of time passed
since a prayer time, I've splitted updateNextPrayer into 2. One for
updating the nextPrayerIndex, and the other for calculating the time
delta.

The nextPrayerIndex will not be updated unless the remaining time until
the next prayer is <= 30mins, or the page was just opened.

calcTimeDelta was redundant, so I removed it.

Also, the progress bar's width is reduced to 70vw from 80vw. Changes are
expected though.