Refactor updateNextPrayerIndex
This time, I added the functionality to presist the nextPrayerIndex if
the next prayer is under 30 minutes.

In other words, nextPrayerIndex will only get updated if the next prayer
is 30 minutes due.

For dhuha however, nextPrayerIndex will get updated to dhuha (idx 3) if
it is 15 minutes due instead.

Constants are assigned at the top of the file to easily make changes to
the threshold.